\name{examplecovariates}
\alias{examplecovariates}
\title{Example covariates dataset for coxKM.}
\description{Example covariates dataset for coxKM.}

\format{
examplecovariates contains:
\describe{
a numeric matrix of 2000 individuals and 2 covariates. Each row represents a different individual. coxKM.examplecovariates is identical to X in SKAT.example.
}

}



\author{Xinyi (Cindy) Lin}





