\name{examplephenotype1}
\alias{examplephenotype1}
\title{Example phenotype for coxKM.}
\description{Example phenotype for coxKM.}

\format{
examplephenotype1 contains:
\describe{
a numeric matrix with two columns, the first column is the event time, the second column is the status indicator. Each row represents a different individual.
}

}



\author{Xinyi (Cindy) Lin}





