\name{examplephenotype2}
\alias{examplephenotype2}
\title{Example phenotype for coxKM.}
\description{Example phenotype for coxKM.}

\format{
examplephenotype2 contains:
\describe{
a numeric matrix with two columns, the first column is the event time, the second column is the status indicator. Each row represents a different individual.
}

}



\author{Xinyi (Cindy) Lin}





