\name{examplesnpset}
\alias{examplesnpset}
\title{Example SNP-set for coxKM.}
\description{Example SNP-set for coxKM.}

\format{
examplesnpset contains:
\describe{
a numeric genotype matrix of 2000 individuals and 11 SNPs. Each row represents a different individual, and each column represents a different SNP marker. coxKM.examplesnpset is subset of Z in SKAT.example.
}
}



\author{Xinyi (Cindy) Lin}





